update/ move top search bar to modules search
