+++ common/components/search/GroupCard/index.tsx
