+++ common/components/search/StoreCard/data.tsx
