+++ common/components/search/StoreCard/index.tsx
